requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
aiosqlite==0.19.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Async Support
aiofiles==23.2.1
aiohttp==3.9.1
asyncio-mqtt==0.16.1

# Voice/Audio
speechrecognition==3.10.0
pyttsx3==2.90
pyaudio==0.2.11
webrtcvad==2.0.10

# Device Communication
paho-mqtt==1.6.1
bleak==0.20.2
pyserial==3.5

# Web Interface
jinja2==3.1.2
python-socketio==5.10.0

# Monitoring & Logging
structlog==23.2.0
prometheus-client==0.19.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
mypy==1.7.1
